initialize gifts state
