--- src/components/Gifts.js
@@ -1,16 +1,0 @@
-function Gifts() {
-  const giftArrayStyles = {
-    border: "1px solid black",
-    // width: "500px",
-    padding: "25px",
-    // margin: "0 auto",
-  };
-  return (
-    <div style={giftArrayStyles}>
-      <h1>gifts component</h1>
-      <p>put your array of gift objects here</p>
-    </div>
-  );
-}
-
-export default Gifts;
